+++ stac-example/src/main/scala/geotrellis/stac/api/SearchFilters.scala
@@ -29,8 +29,12 @@ import cats.Semigroup
 import cats.syntax.option._
 import cats.syntax.either._
 import cats.syntax.semigroup._
+import cats.syntax.apply._
 import cats.instances.option._
+import cats.instances.either._
 import cats.instances.list._
+
+import java.time.Instant
 
 case class SearchFilters(
   bbox: Option[Bbox] = None,
@@ -44,6 +48,53 @@ case class SearchFilters(
 )
 
 object SearchFilters {
+  // TemporalExtent STAC API compatible serialization
+  private def stringToInstant(s: String): Either[Throwable, Instant] =
+    Either.catchNonFatal(Instant.parse(s))
+
+  private def temporalExtentToString(te: TemporalExtent): String =
+    te.value match {
+      case Some(start) :: Some(end) :: _ if start != end => s"${start.toString}/${end.toString}"
+      case Some(start) :: Some(end) :: _ if start == end => s"${start.toString}"
+      case Some(start) :: None :: _                      => s"${start.toString}/.."
+      case None :: Some(end) :: _                        => s"../${end.toString}"
+    }
+
+  private def temporalExtentFromString(str: String): Either[String, TemporalExtent] = {
+    str.split("/").toList match {
+      case ".." :: endString :: _        =>
+        val parsedEnd = stringToInstant(endString)
+        parsedEnd match {
+          case Left(_)             => s"Could not decode instant: $str".asLeft
+          case Right(end: Instant) => TemporalExtent(None, end).asRight
+        }
+      case startString :: ".." :: _      =>
+        val parsedStart = stringToInstant(startString)
+        parsedStart match {
+          case Left(_)               => s"Could not decode instant: $str".asLeft
+          case Right(start: Instant) => TemporalExtent(start, None).asRight
+        }
+      case startString :: endString :: _ =>
+        val parsedStart = stringToInstant(startString)
+        val parsedEnd   = stringToInstant(endString)
+        (parsedStart, parsedEnd).tupled match {
+          case Left(_)                               => s"Could not decode instant: $str".asLeft
+          case Right((start: Instant, end: Instant)) => TemporalExtent(start, end).asRight
+        }
+      case _                             =>
+        Either.catchNonFatal(Instant.parse(str)) match {
+          case Left(_)           => s"Could not decode instant: $str".asLeft
+          case Right(t: Instant) => TemporalExtent(t, t).asRight
+        }
+    }
+  }
+
+  implicit val encoderTemporalExtent: Encoder[TemporalExtent] =
+    Encoder.encodeString.contramap[TemporalExtent](temporalExtentToString)
+
+  implicit val decoderTemporalExtent: Decoder[TemporalExtent] =
+    Decoder.decodeString.emap(temporalExtentFromString)
+
   object IntersectionSemigroup {
     implicit val bboxIntersectionSemigroup: Semigroup[Bbox] = { (left, right) =>
       val Extent(xmin, ymin, xmax, ymax) =